Move session.py under clients/db
