+++ app/api/api_v1/routers/documents.py
@@ -8,12 +8,12 @@ from app.api.api_v1.schemas.document import (
     FamilyAndDocumentsResponse,
     FamilyDocumentWithContextResponse,
 )
+from app.clients.db.session import get_db
 from app.db.crud.document import (
     get_family_and_documents,
     get_family_document_and_context,
     get_slugged_objects,
 )
-from app.db.session import get_db
 
 _LOGGER = logging.getLogger(__file__)
