+++ app/api/api_v1/routers/lookups/geo_stats.py
@@ -8,7 +8,7 @@ from pydantic import BaseModel
 from sqlalchemy import exc
 
 from app.api.api_v1.routers.lookups.router import lookups_router
-from app.db.session import get_db
+from app.clients.db.session import get_db
 
 _LOGGER = logging.getLogger(__name__)
