+++ app/api/api_v1/routers/search.py
@@ -18,6 +18,7 @@ from sqlalchemy.orm import Session
 from starlette.responses import RedirectResponse
 
 from app.api.api_v1.schemas.search import SearchRequestBody, SearchResponse
+from app.clients.db.session import get_db
 from app.core.aws import S3Client, S3Document, get_s3_client
 from app.core.config import (
     AWS_REGION,
@@ -36,7 +37,6 @@ from app.core.search import (
     process_result_into_csv,
     process_vespa_search_response,
 )
-from app.db.session import get_db
 
 _LOGGER = logging.getLogger(__name__)
