+++ tests/conftest.py
@@ -15,10 +15,10 @@ from sqlalchemy.engine import Engine
 from sqlalchemy.orm import sessionmaker
 from sqlalchemy_utils import create_database, database_exists, drop_database
 
+from app.clients.db.session import get_db
 from app.core import custom_app, security
 from app.core.aws import S3Client, get_s3_client
 from app.core.custom_app import AppTokenFactory
-from app.db.session import get_db
 from app.main import app
