+++ tests/non_search/app/lookups/test_config.py
@@ -11,8 +11,8 @@ from db_client.models.dfce.family import (
 )
 from db_client.models.organisation import Corpus, Organisation
 
+from app.clients.db.session import SessionLocal
 from app.core.util import tree_table_to_json
-from app.db.session import SessionLocal
 
 LEN_ORG_CONFIG = 3
 EXPECTED_CCLW_TAXONOMY = {
